Step 1 - NoRelVal - fix 1

+++ CondFormats/HcalObjects/interface/HcalQIETyp.h
@@ -13,7 +13,7 @@ class HcalQIETyp
 {
  public:
   HcalQIETyp(): mId(0), mValue(0) {}
-  HcalQIETyp(unsigned long fid, float value): mId(fid), mValue(value) {}
+  HcalQIETyp(uint32_t fid, int value): mId(fid), mValue(value) {}
 
   uint32_t rawId() const {return mId;}
